migrating map/list/json utility functions from kafka-common into new module data-common so these can be used in places like elastic-common without requiring unrelated kafka dependencies
